requirements.txt
django>=5.0
djangorestframework
PyMuPDF
openai
python-dotenv
jsonschema
requests
